requirement.txt
appdirs==1.4.3
astroid==2.4.2
Babel==2.9.0
blinker==1.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-WTF==0.14.3
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
limits==1.5.1
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
passlib==1.7.4
pep517==0.8.2
progress==1.5
psycopg2-binary==2.8.6
pylint==2.6.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.5
redis==3.5.3
requests==2.22.0
retrying==1.3.3
secure-smtplib==0.1.1
six==1.14.0
speaklater==1.3
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
